sk-api-integration/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking
#
aiohttp==3.8.6
    # via microsoft-kiota-authentication-azure
aiosignal==1.3.1
    # via aiohttp
anyio==4.0.0
    # via httpcore
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
azure-core==1.29.5
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.15.0
    # via
    #   -r requirements.in
    #   msgraph-sdk
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
cryptography==41.0.5
    # via
    #   azure-identity
    #   msal
    #   pyjwt
deprecated==1.2.14
    # via opentelemetry-api
exceptiongroup==1.1.3
    # via anyio
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.18.0
    # via httpx
httpx[http2]==0.25.0
    # via
    #   httpx
    #   microsoft-kiota-http
    #   msgraph-core
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via opentelemetry-api
microsoft-kiota-abstractions==1.0.0
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-text
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-authentication-azure==1.0.0
    # via msgraph-sdk
microsoft-kiota-http==1.0.0
    # via
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-serialization-json==1.0.0
    # via msgraph-sdk
microsoft-kiota-serialization-text==1.0.0
    # via msgraph-sdk
msal==1.24.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
msgraph-core==1.0.0a4
    # via msgraph-sdk
msgraph-sdk==1.0.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.20.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
opentelemetry-sdk==1.20.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.41b0
    # via opentelemetry-sdk
pendulum==2.1.2
    # via microsoft-kiota-serialization-json
portalocker==2.8.2
    # via msal-extensions
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
python-dateutil==2.8.2
    # via
    #   microsoft-kiota-serialization-text
    #   pendulum
pytzdata==2020.1
    # via pendulum
requests==2.31.0
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
std-uritemplate==0.0.46
    # via microsoft-kiota-abstractions
typing-extensions==4.8.0
    # via
    #   azure-core
    #   opentelemetry-sdk
urllib3==2.0.7
    # via requests
wrapt==1.15.0
    # via deprecated
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata
    
flask==3.0.0
plaid-python==21.0.0
msal==1.24.1
requests==2.31.0
python-dotenv==1.0.0
